+++ src/matrix4x4_math.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::inline_always)]
 #![allow(unused)]
 
-use crate::{Matrix3x3, Matrix3x3Math, Matrix4x4, Vector4d};
+use crate::{Matrix3x3, Matrix3x3Math, Matrix4x4, Quaternion, Vector4d};
 
 // **** Math ****
 
@@ -16,6 +16,7 @@ pub trait Matrix4x4Math: Sized {
     fn m4x4_mul_vector(this: Matrix4x4<Self>, other: Vector4d<Self>) -> Vector4d<Self>;
     fn m4x4_vector_mul(this: Vector4d<Self>, other: Matrix4x4<Self>) -> Vector4d<Self>;
     fn m4x4_vector_outer_product(col: Vector4d<Self>, row: Vector4d<Self>) -> Matrix4x4<Self>;
+    fn m4x4_quaternion_outer_product(this: Quaternion<Self>) -> Matrix4x4<Self>;
     fn m4x4_mul(this: Matrix4x4<Self>, other: Matrix4x4<Self>) -> Matrix4x4<Self>;
     fn m4x4_determinant(this: Matrix4x4<Self>) -> Self;
     fn m4x4_top_right_determinant(this: Matrix4x4<Self>) -> Self;
@@ -93,39 +94,73 @@ impl Matrix4x4Math for f32 {
         }
     }
 
-    #[rustfmt::skip]
     #[inline(always)]
     fn m4x4_vector_outer_product(col: Vector4d<Self>, row: Vector4d<Self>) -> Matrix4x4<Self> {
         // Structure data into local fixed-size arrays of 4 elements.
         // Since row is align(16), we manually map the implicit 4th buffer element.
         let r = [row.x, row.y, row.z, row.t];
 
+        let mut m0 = [0.0; 4];
         let mut m1 = [0.0; 4];
         let mut m2 = [0.0; 4];
         let mut m3 = [0.0; 4];
-        let mut m4 = [0.0; 4];
 
         // Write uniform loops spanning exactly 4 elements.
         // LLVM's auto-vectorizer recognizes 4-wide float operations
         // and combines these into parallel execution blocks, if the processor supports it.
         for ii in 0..4 {
-            m1[ii] = col.x * r[ii];
+            m0[ii] = col.x * r[ii];
         }
         for ii in 0..4 {
-            m2[ii] = col.y * r[ii];
+            m1[ii] = col.y * r[ii];
         }
         for ii in 0..4 {
-            m3[ii] = col.z * r[ii];
+            m2[ii] = col.z * r[ii];
         }
         for ii in 0..4 {
-            m4[ii] = col.t * r[ii];
+            m3[ii] = col.t * r[ii];
         }
 
         Matrix4x4::from([
-            m1[0], m1[1], m1[2], m1[3],
-            m2[0], m2[1], m2[2], m2[3],
-            m3[0], m3[1], m3[2], m3[3],
-            m3[0], m3[1], m3[2], m4[3],
+            m0[0], m0[1], m0[2], m0[3], //
+            m1[0], m1[1], m1[2], m1[3], //
+            m2[0], m2[1], m2[2], m2[3], //
+            m3[0], m3[1], m3[2], m3[3], //
+        ])
+    }
+
+    #[inline(always)]
+    fn m4x4_quaternion_outer_product(this: Quaternion<Self>) -> Matrix4x4<Self> {
+        // Structure data into local fixed-size arrays of 4 elements.
+        // Since row is align(16), we manually map the implicit 4th buffer element.
+        let r = [this.w, this.x, this.y, this.z];
+
+        let mut m0 = [0.0; 4];
+        let mut m1 = [0.0; 4];
+        let mut m2 = [0.0; 4];
+        let mut m3 = [0.0; 4];
+
+        // Write uniform loops spanning exactly 4 elements.
+        // LLVM's auto-vectorizer recognizes 4-wide float operations
+        // and combines these into parallel execution blocks, if the processor supports it.
+        for ii in 0..4 {
+            m0[ii] = this.w * r[ii];
+        }
+        for ii in 0..4 {
+            m1[ii] = this.x * r[ii];
+        }
+        for ii in 0..4 {
+            m2[ii] = this.y * r[ii];
+        }
+        for ii in 0..4 {
+            m3[ii] = this.z * r[ii];
+        }
+
+        Matrix4x4::from([
+            m0[0], m0[1], m0[2], m0[3], //
+            m1[0], m1[1], m1[2], m1[3], //
+            m2[0], m2[1], m2[2], m2[3], //
+            m3[0], m3[1], m3[2], m3[3], //
         ])
     }
 
@@ -310,39 +345,73 @@ impl Matrix4x4Math for f64 {
         }
     }
 
-    #[rustfmt::skip]
     #[inline(always)]
     fn m4x4_vector_outer_product(col: Vector4d<Self>, row: Vector4d<Self>) -> Matrix4x4<Self> {
         // Structure data into local fixed-size arrays of 4 elements.
         // Since row is align(16), we manually map the implicit 4th buffer element.
         let r = [row.x, row.y, row.z, row.t];
 
+        let mut m0 = [0.0; 4];
         let mut m1 = [0.0; 4];
         let mut m2 = [0.0; 4];
         let mut m3 = [0.0; 4];
-        let mut m4 = [0.0; 4];
 
         // Write uniform loops spanning exactly 4 elements.
         // LLVM's auto-vectorizer recognizes 4-wide float operations
         // and combines these into parallel execution blocks, if the processor supports it.
         for ii in 0..4 {
-            m1[ii] = col.x * r[ii];
+            m0[ii] = col.x * r[ii];
         }
         for ii in 0..4 {
-            m2[ii] = col.y * r[ii];
+            m1[ii] = col.y * r[ii];
         }
         for ii in 0..4 {
-            m3[ii] = col.z * r[ii];
+            m2[ii] = col.z * r[ii];
         }
         for ii in 0..4 {
-            m4[ii] = col.t * r[ii];
+            m3[ii] = col.t * r[ii];
         }
 
         Matrix4x4::from([
-            m1[0], m1[1], m1[2], m1[3],
-            m2[0], m2[1], m2[2], m2[3],
-            m3[0], m3[1], m3[2], m3[3],
-            m3[0], m3[1], m3[2], m4[3],
+            m0[0], m0[1], m0[2], m0[3], //
+            m1[0], m1[1], m1[2], m1[3], //
+            m2[0], m2[1], m2[2], m2[3], //
+            m3[0], m3[1], m3[2], m3[3], //
+        ])
+    }
+
+    #[inline(always)]
+    fn m4x4_quaternion_outer_product(this: Quaternion<Self>) -> Matrix4x4<Self> {
+        // Structure data into local fixed-size arrays of 4 elements.
+        // Since row is align(16), we manually map the implicit 4th buffer element.
+        let r = [this.w, this.x, this.y, this.z];
+
+        let mut m0 = [0.0; 4];
+        let mut m1 = [0.0; 4];
+        let mut m2 = [0.0; 4];
+        let mut m3 = [0.0; 4];
+
+        // Write uniform loops spanning exactly 4 elements.
+        // LLVM's auto-vectorizer recognizes 4-wide float operations
+        // and combines these into parallel execution blocks, if the processor supports it.
+        for ii in 0..4 {
+            m0[ii] = this.w * r[ii];
+        }
+        for ii in 0..4 {
+            m1[ii] = this.x * r[ii];
+        }
+        for ii in 0..4 {
+            m2[ii] = this.y * r[ii];
+        }
+        for ii in 0..4 {
+            m3[ii] = this.z * r[ii];
+        }
+
+        Matrix4x4::from([
+            m0[0], m0[1], m0[2], m0[3], //
+            m1[0], m1[1], m1[2], m1[3], //
+            m2[0], m2[1], m2[2], m2[3], //
+            m3[0], m3[1], m3[2], m3[3], //
         ])
     }