+++ src/vector2d.rs
@@ -494,6 +494,30 @@ impl<T> IndexMut<usize> for Vector2d<T> {
     }
 }
 
+// **** lerp ****
+
+impl<T> Vector2d<T>
+where
+    T: Copy,
+    Vector2d<T>: Mul<T, Output = Vector2d<T>> + Add<Output = Vector2d<T>> + Sub<Output = Vector2d<T>>,
+{
+    /// Linear interpolation between two vectors.
+    /// Calculates `self * (1 - t) + other * t`.
+    /// ```
+    /// # use vqm::Vector2df32;
+    /// let u = Vector2df32::new(2.0, 5.0);
+    /// let v = Vector2df32::new(3.0, 7.0);
+    /// let w = u.lerp(v, 0.25);
+    ///
+    /// assert_eq!(w, Vector2df32::new(2.25, 5.5));
+    /// ```
+    #[inline]
+    #[must_use]
+    pub fn lerp(self, other: Self, t: T) -> Self {
+        self + (other - self) * t
+    }
+}
+
 // **** abs ****
 
 impl<T> Vector2d<T>